+++ noodles-sam/src/lazy/record/quality_scores.rs
@@ -1,0 +1,25 @@
+/// Raw SAM record quality scores.
+#[derive(Debug, Eq, PartialEq)]
+pub struct QualityScores<'a>(&'a [u8]);
+
+impl<'a> QualityScores<'a> {
+    pub(super) fn new(buf: &'a [u8]) -> Self {
+        Self(buf)
+    }
+
+    /// Returns whether there are any scores.
+    pub fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+
+    /// Returns the number of scores.
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+}
+
+impl<'a> AsRef<[u8]> for QualityScores<'a> {
+    fn as_ref(&self) -> &[u8] {
+        self.0
+    }
+}